% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_coherence.R
\name{coherence}
\alias{coherence}
\title{Helper function for calculating coherence for a single topic's worth of terms}
\usage{
coherence(dtm_data, top_terms, smoothing_beta)
}
\arguments{
\item{dtm_data}{a document-term matrix of token counts coercible to \code{simple_triplet_matrix}}

\item{top_terms}{a character vector of the top terms for a given topic}

\item{smoothing_beta}{a numeric indicating the value to use to smooth the document frequencies
in order avoid log zero issues, the default is 1}
}
\value{
a numeric indicating coherence for the topic
}
\description{
Helper function for calculating coherence for a single topic's worth of terms
}
