% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_diagnostics.R
\name{contain_equal_docs}
\alias{contain_equal_docs}
\title{Helper function to check that a topic model and a dtm contain the same number of documents}
\usage{
contain_equal_docs(topic_model, dtm_data)
}
\arguments{
\item{topic_model}{a fitted topic model object from one of the following:
\code{\link[topicmodels]{tm-class}}}

\item{dtm_data}{a document-term matrix of token counts coercible to \code{simple_triplet_matrix}}
}
\value{
a logical indicating whether or not the two object contain the same number of documents
}
\description{
Helper function to check that a topic model and a dtm contain the same number of documents
}
