% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_diagnostics.R
\name{n_topics}
\alias{n_topics}
\title{Helper function to determine the number of topics in a topic model}
\usage{
n_topics(topic_model)
}
\arguments{
\item{topic_model}{a fitted topic model object from one of the following:
\code{\link[topicmodels]{LDA}}, \code{\link[topicmodels]{CTM}}}
}
\value{
an integer indicating the number of topics in the topic model
}
\description{
Helper function to determine the number of topics in a topic model
}
