% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_df_dist.R
\name{tf_df_dist_diff}
\alias{tf_df_dist_diff}
\title{Helper function to calculate the Hellinger distance
between the token frequencies and document frequencies
for a specific topic's top N tokens}
\usage{
tf_df_dist_diff(dtm, top_terms)
}
\arguments{
\item{dtm}{- the document-term matrix coericble to \code{simple_triplet_matrix}}

\item{top_terms}{- a character vector of the top N tokens}
}
\value{
a single value representing the Hellinger distance
}
\description{
Helper function to calculate the Hellinger distance
between the token frequencies and document frequencies
for a specific topic's top N tokens
}
